requirements.txt
django >= 1.8
django-material >= 0.2.0
django-autoslug >= 1.7.2
django-admin-sortable >= 1.8.3
django-mptt >= 0.7.1
django-polymorphic >= 0.7
django-filer >= 0.9.9
django-compressor >= 1.4
